feat: get connected, styling
