+++ src/components/pages/Home/FeatureSummaryItem.tsx
@@ -10,16 +10,18 @@ interface FeaturesSummaryItemProps {
 }
 
 const FeaturesSummaryItem: FC<FeaturesSummaryItemProps> = ({ imgSrc, heading, subHeading }) => (
-  <Row className="d-flex flex-row">
-    <Col xs={3} className="text-end mt-2">
-      <Image src={imgSrc} width="64px" loading="lazy" alt="Feature image" />
-    </Col>
+  <Col className="pe-3">
+    <Row className="d-flex flex-row">
+      <Col xs={3} className="text-end mt-2">
+        <Image src={imgSrc} width="64px" loading="lazy" alt="Feature image" />
+      </Col>
 
-    <Col xs={9} className="text-start">
-      <h4>{heading}</h4>
-      <p>{subHeading}</p>
-    </Col>
-  </Row>
+      <Col xs={9} className="text-start">
+        <h4>{heading}</h4>
+        <p>{subHeading}</p>
+      </Col>
+    </Row>
+  </Col>
 )
 
 export default FeaturesSummaryItem
