+++ src/components/pages/Home/Features.tsx
@@ -6,8 +6,8 @@ import SupportImg from '@/assets/images/support.jpg'
 import FeaturesItem from './FeaturesItem'
 
 const Features: FC = () => (
-  <section className="section-light features pt-5 pb-4">
-    <Container className="d-grid gap-3">
+  <section className="section-light features">
+    <Container>
       <FeaturesItem
         imgSrc={SpeedImg as string}
         imageOrder={1}
@@ -15,7 +15,6 @@ const Features: FC = () => (
         heading="Accelerate to your goals"
         subHeading="No more waiting for videos to load or games to buffer. Experience lightning-fast internet speeds that allow you to download files in seconds, stream 4K content without interruptions, and enjoy lag-free online gaming."
       />
-      <div className="mt-lg-3" />
       <FeaturesItem
         imgSrc={WifiImg as string}
         imageOrder={2}
@@ -23,7 +22,6 @@ const Features: FC = () => (
         heading="Stay connected"
         subHeading="Tired of dead zones and unreliable connections? Our advanced routers deliver powerful Wi-Fi signals that reach every corner of your home, ensuring a seamless connection for all your devices. Browse, stream, and game anywhere in your house with confidence."
       />
-      <div className="mt-lg-3" />
       <FeaturesItem
         imgSrc={SupportImg as string}
         imageOrder={1}
