+++ src/components/pages/Home/FeaturesSummaryMore.tsx
@@ -8,24 +8,22 @@ import WorkImg from '@/assets/icons/working.png'
 import FeaturesSummaryItem from './FeatureSummaryItem'
 
 const FeaturesSummaryMore: FC = () => (
-  <section className="section-dark" id="features-summary">
+  <section className="section-dark features-summary">
     <Container>
       <Row className="">
         <Col className="text-center my-4">
           <h1 className="primary">More Than Just Speed</h1>
-          <div className="d-flex flex-column flex-lg-row mt-4">
+          <div className="item-container">
             <FeaturesSummaryItem
               imgSrc={FamilyImg as string}
               heading="Perfect for Families"
               subHeading="Connect everyone with unlimited data plans for uninterrupted entertainment."
             />
-            <div className="me-lg-3 mb-3" />
             <FeaturesSummaryItem
               imgSrc={GamingImg as string}
               heading="Unleash the Gamer Within"
               subHeading="Experience low latency and smooth gameplay for ultimate victory."
             />
-            <div className="me-lg-3 mb-3" />
             <FeaturesSummaryItem
               imgSrc={WorkImg as string}
               heading="Work From Anywhere"
